requirements.txt
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
elasticsearch==7.17.1
pyyaml==6.0.1
schedule==1.2.1
pyinotify==0.9.6
python-iptables==1.0.0
